+++ src/api/Methods.ts
@@ -1,4 +1,13 @@
-import { AppIntent, Channel, Context, ContextHandler, IntentResolution, Listener, ImplementationMetadata } from '..';
+import {
+  AppIntent,
+  Channel,
+  Context,
+  ContextHandler,
+  IntentHandler,
+  IntentResolution,
+  Listener,
+  ImplementationMetadata,
+} from '..';
 import { TargetApp } from './Types';
 
 const DEFAULT_TIMEOUT = 5000;
@@ -51,12 +60,12 @@ export function open(app: TargetApp, context?: Context): Promise<void> {
   return rejectIfNoGlobal(() => window.fdc3.open(app, context));
 }
 
-export function findIntent(intent: string, context?: Context): Promise<AppIntent> {
-  return rejectIfNoGlobal(() => window.fdc3.findIntent(intent, context));
+export function findIntent(intent: string, context?: Context, resultType?: string): Promise<AppIntent> {
+  return rejectIfNoGlobal(() => window.fdc3.findIntent(intent, context, resultType));
 }
 
-export function findIntentsByContext(context: Context): Promise<AppIntent[]> {
-  return rejectIfNoGlobal(() => window.fdc3.findIntentsByContext(context));
+export function findIntentsByContext(context: Context, resultType?: string): Promise<AppIntent[]> {
+  return rejectIfNoGlobal(() => window.fdc3.findIntentsByContext(context, resultType));
 }
 
 export function broadcast(context: Context): Promise<void> {
@@ -71,7 +80,7 @@ export function raiseIntentForContext(context: Context, app?: TargetApp): Promis
   return rejectIfNoGlobal(() => window.fdc3.raiseIntentForContext(context, app));
 }
 
-export function addIntentListener(intent: string, handler: ContextHandler): Promise<Listener> {
+export function addIntentListener(intent: string, handler: IntentHandler): Promise<Listener> {
   return rejectIfNoGlobal(() => window.fdc3.addIntentListener(intent, handler));
 }
 
@@ -87,12 +96,36 @@ export function addContextListener(
   }
 }
 
+export function getUserChannels(): Promise<Channel[]> {
+  return rejectIfNoGlobal(() => {
+    //fallback to getSystemChannels for FDC3 <2.0 implementations
+    if (window.fdc3.getUserChannels) {
+      return window.fdc3.getUserChannels();
+    } else {
+      return window.fdc3.getSystemChannels();
+    }
+  });
+}
+
 export function getSystemChannels(): Promise<Channel[]> {
-  return rejectIfNoGlobal(() => window.fdc3.getSystemChannels());
+  //fallforward to getUserChannels for FDC3 2.0+ implementations
+  return getUserChannels();
+}
+
+export function joinUserChannel(channelId: string): Promise<void> {
+  return rejectIfNoGlobal(() => {
+    //fallback to joinChannel for FDC3 <2.0 implementations
+    if (window.fdc3.joinUserChannel) {
+      return window.fdc3.joinUserChannel(channelId);
+    } else {
+      return window.fdc3.joinChannel(channelId);
+    }
+  });
 }
 
 export function joinChannel(channelId: string): Promise<void> {
-  return rejectIfNoGlobal(() => window.fdc3.joinChannel(channelId));
+  //fallforward to joinUserChannel for FDC3 2.0+ implementations
+  return joinUserChannel(channelId);
 }
 
 export function getOrCreateChannel(channelId: string): Promise<Channel> {