+++ test/Methods.test.ts
@@ -13,9 +13,11 @@ import {
   getCurrentChannel,
   getInfo,
   getOrCreateChannel,
+  getUserChannels,
   getSystemChannels,
   ImplementationMetadata,
   joinChannel,
+  joinUserChannel,
   leaveCurrentChannel,
   open,
   raiseIntent,
@@ -41,6 +43,10 @@ expect.extend({
     expect(received).toThrowError(UnavailableError);
     return { pass: true } as jest.CustomMatcherResult;
   },
+});
+
+beforeEach(() => {
+  jest.resetAllMocks();
 });
 
 describe('test ES6 module', () => {
@@ -79,12 +85,16 @@ describe('test ES6 module', () => {
       await expect(addContextListener(expect.any(String), expect.any(Object))).rejects.toEqual(UnavailableError);
     });
 
-    test('getSystemChannels should reject', async () => {
-      await expect(getSystemChannels()).rejects.toEqual(UnavailableError);
+    test('getUserChannels should reject', async () => {
+      await expect(getUserChannels()).rejects.toEqual(UnavailableError);
     });
 
     test('joinChannel should reject', async () => {
       await expect(joinChannel(expect.any(String))).rejects.toEqual(UnavailableError);
+    });
+
+    test('joinUserChannel should reject', async () => {
+      await expect(joinUserChannel(expect.any(String))).rejects.toEqual(UnavailableError);
     });
 
     test('getOrCreateChannel should reject', async () => {
@@ -128,14 +138,30 @@ describe('test ES6 module', () => {
       await findIntent(intent, ContactContext);
 
       expect(window.fdc3.findIntent).toHaveBeenCalledTimes(1);
-      expect(window.fdc3.findIntent).toHaveBeenCalledWith(intent, ContactContext);
+      expect(window.fdc3.findIntent).toHaveBeenLastCalledWith(intent, ContactContext, undefined);
+    });
+
+    test('findIntent should delegate to window.fdc3.findIntent (with additional output type argument)', async () => {
+      const intent = 'ViewChart';
+
+      await findIntent(intent, ContactContext, ContextTypes.Contact);
+
+      expect(window.fdc3.findIntent).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.findIntent).toHaveBeenLastCalledWith(intent, ContactContext, ContextTypes.Contact);
     });
 
     test('findIntentsByContext should delegate to window.fdc3.findIntentsByContext', async () => {
       await findIntentsByContext(ContactContext);
 
       expect(window.fdc3.findIntentsByContext).toHaveBeenCalledTimes(1);
-      expect(window.fdc3.findIntentsByContext).toHaveBeenCalledWith(ContactContext);
+      expect(window.fdc3.findIntentsByContext).toHaveBeenLastCalledWith(ContactContext, undefined);
+    });
+
+    test('findIntentsByContext should delegate to window.fdc3.findIntentsByContext (with additional output type argument)', async () => {
+      await findIntentsByContext(ContactContext, ContextTypes.Contact);
+
+      expect(window.fdc3.findIntentsByContext).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.findIntentsByContext).toHaveBeenLastCalledWith(ContactContext, ContextTypes.Contact);
     });
 
     test('broadcast should delegate to window.fdc3.broadcast', async () => {
@@ -187,20 +213,36 @@ describe('test ES6 module', () => {
       expect(window.fdc3.addContextListener).toHaveBeenNthCalledWith(2, null, handler2);
     });
 
-    test('getSystemChannels should delegate to window.fdc3.getSystemChannels', async () => {
-      await getSystemChannels();
+    test('getUserChannels should delegate to window.fdc3.getUserChannels', async () => {
+      await getUserChannels();
 
-      expect(window.fdc3.getSystemChannels).toHaveBeenCalledTimes(1);
-      expect(window.fdc3.getSystemChannels).toHaveBeenCalledWith();
+      expect(window.fdc3.getUserChannels).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.getUserChannels).toHaveBeenCalledWith();
     });
 
-    test('joinChannel should delegate to window.fdc3.joinChannel', async () => {
+    test('getSystemChannels should delegate to window.fdc3.getUserChannels', async () => {
+      await getSystemChannels();
+
+      expect(window.fdc3.getUserChannels).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.getUserChannels).toHaveBeenCalledWith();
+    });
+
+    test('joinChannel should delegate to window.fdc3.joinUserChannel', async () => {
       const channelId = 'channel';
 
       await joinChannel(channelId);
 
-      expect(window.fdc3.joinChannel).toHaveBeenCalledTimes(1);
-      expect(window.fdc3.joinChannel).toHaveBeenCalledWith(channelId);
+      expect(window.fdc3.joinUserChannel).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.joinUserChannel).toHaveBeenCalledWith(channelId);
+    });
+
+    test('joinUserChannel should delegate to window.fdc3.joinUserChannel', async () => {
+      const channelId = 'channel';
+
+      await joinUserChannel(channelId);
+
+      expect(window.fdc3.joinUserChannel).toHaveBeenCalledTimes(1);
+      expect(window.fdc3.joinUserChannel).toHaveBeenCalledWith(channelId);
     });
 
     test('getOrCreateChannel should delegate to window.fdc3.getOrCreateChannel', async () => {